+++ nuxt/utils/util.ts
@@ -2,22 +2,20 @@ import { IncomingMessage } from 'node:http'
 
 import { CombinedError } from '@urql/core'
 import Clipboard from 'clipboard'
-import { GraphQLError } from 'graphql'
+import consola from 'consola'
+import { defu } from 'defu'
+import { H3Event, getCookie } from 'h3'
+import { ofetch } from 'ofetch'
 import Swal from 'sweetalert2'
-import { ComputedRef, Ref } from 'vue'
+import { Ref } from 'vue'
 import { LocationQueryValue } from 'vue-router'
 
-import { REGEX_UUID } from './constants'
-
-export type BackendError = {
-  graphQLErrors: (GraphQLError & { originalError?: { errcode?: string } })[]
-} & CombinedError
-
-export type ApiData = ComputedRef<{
-  data?: Object
-  errors: BackendError[]
-  isFetching: boolean
-}>
+import { TIMEZONE_COOKIE_NAME, REGEX_UUID } from './constants'
+import type {
+  ArrayElement,
+  BackendError,
+  UnionToIntersection,
+} from '~/types/types'
 
 export function append(path: string, pathToAppend: string): string {
   return path + (path.endsWith('/') ? '' : '/') + pathToAppend
@@ -49,6 +47,79 @@ export function copyText(text: string) {
     })
     fakeElement.click()
   })
+}
+
+function getCsp(host: string): Record<string, Array<string>> {
+  const hostName = host.replace(/:[0-9]+$/, '')
+  const config = useRuntimeConfig()
+
+  const stagingHostOrHost = config.public.stagingHost || host
+
+  const base = {
+    'base-uri': ["'none'"], // Mozilla Observatory.
+    'connect-src': [
+      "'self'",
+      'blob:', // vue-advanced-cropper
+      `https://${getDomainTldPort(stagingHostOrHost)}`, // `/api` requests
+      `https://postgraphile.${getDomainTldPort(stagingHostOrHost)}`, // backend requests
+      `https://tusd.${getDomainTldPort(stagingHostOrHost)}`, // image upload requests
+      'https://*.google-analytics.com', // Google Analytics 4 (https://developers.google.com/tag-platform/tag-manager/web/csp)
+      'https://*.analytics.google.com', // Google Analytics 4 (https://developers.google.com/tag-platform/tag-manager/web/csp)
+    ],
+    'default-src': ["'none'"],
+    'font-src': ["'self'"], // ~/public/assets/static/fonts
+    'form-action': ["'self'"], // Mozilla Observatory: "none".
+    'frame-ancestors': ["'none'"], // Mozilla Observatory.
+    'frame-src': ['https://challenges.cloudflare.com'],
+    'img-src': [
+      "'self'",
+      'blob:',
+      'data:',
+      `https://tusd.${getDomainTldPort(stagingHostOrHost)}`,
+      'https://*.google-analytics.com', // Google Analytics 4 (https://developers.google.com/tag-platform/tag-manager/web/csp)
+      'https://www.gravatar.com/avatar/', // profile picture fallback
+    ],
+    'manifest-src': ["'self'"],
+    'media-src': ["'none'"],
+    'object-src': ["'none'"],
+    'prefetch-src': ["'self'"],
+    'report-uri': ['https://dargmuesli.report-uri.com/r/d/csp/enforce'],
+    // TODO: evaluate header (https://github.com/maevsi/maevsi/issues/830) // https://stackoverflow.com/questions/62081028/this-document-requires-trustedscripturl-assignment
+    // 'require-trusted-types-for': ["'script'"], // csp-evaluator
+    'script-src': [
+      'blob:',
+      "'self'",
+      'https://challenges.cloudflare.com/turnstile/v0/api.js', // Cloudfalse Turnstile
+      'https://static.cloudflareinsights.com', // Cloudflare analytics
+      'https://*.google-analytics.com', // Google Analytics 4 (https://developers.google.com/tag-platform/tag-manager/web/csp)
+      'https://www.googletagmanager.com/gtag/js', // Google Analytics 4 (https://developers.google.com/tag-platform/tag-manager/web/csp)
+      "'unsafe-inline'", // https://github.com/unjs/nitro/issues/81
+      "'unsafe-eval'", // https://github.com/unjs/nitro/issues/81
+    ],
+    'style-src': ["'self'", "'unsafe-inline'"], // Tailwind
+  }
+
+  const development = {
+    'connect-src': [
+      `http://${hostName}:24678/_nuxt/`,
+      `https://${hostName}:24678/_nuxt/`,
+      `ws://${hostName}:24678/_nuxt/`,
+      `wss://${hostName}:24678/_nuxt/`,
+    ],
+  }
+
+  const production = {
+    'connect-src': [`https://${stagingHostOrHost}/cdn-cgi/rum`],
+  }
+
+  return defu(base, config.public.isInProduction ? production : development)
+}
+
+export function getCspAsString(event: H3Event): string {
+  const host = getHost(event.node.req)
+  const csp = getCsp(host)
+
+  return Object.keys(csp).reduce((p, c) => `${p}${c} ${csp[c].join(' ')};`, '')
 }
 
 // export function getDeferredPromise<T>(then?: (value: any) => T): Promise<T> {
@@ -85,33 +156,41 @@ export function getHost(req: IncomingMessage) {
   return req.headers.host
 }
 
-export const getApiDataDefault = (): ApiData =>
-  computed(() =>
-    reactive({
-      data: undefined,
-      ...getApiMeta(),
-    })
-  )
-
-export const getApiMeta = (
-  queries?: {
+export const getApiData = <
+  S,
+  T extends {
+    data: Ref<S>
     error: Ref<CombinedError | undefined>
     fetching: Ref<boolean>
-  }[]
-) => ({
-  errors: queries
-    ? queries.reduce((p, c) => {
-        if (c.error.value) {
-          return [...p, c.error.value]
-        } else {
-          return p
-        }
-      }, [] as BackendError[])
-    : [],
-  isFetching: queries
-    ? queries.reduce((p, c) => p || c.fetching.value, false)
-    : false,
-})
+  }
+>(
+  queries?: T[]
+) => {
+  const apiData = computed(() => ({
+    data: (queries || []).reduce(
+      (p, c) => ({ ...p, ...c.data.value }),
+      {} as NonNullable<
+        UnionToIntersection<NonNullable<ArrayElement<T[]>['data']['value']>>
+      >
+    ),
+    errors: (queries || []).reduce(
+      (p, c) => (c.error.value ? [...p, c.error.value] : p),
+      [] as BackendError[]
+    ),
+    isFetching: (queries || []).reduce((p, c) => p || c.fetching.value, false),
+  }))
+
+  watch(
+    () => apiData.value.errors,
+    (current, previous) => {
+      current
+        .filter((error) => !previous.includes(error))
+        .forEach((error) => consola.error(error))
+    }
+  )
+
+  return apiData
+}
 
 export function getCombinedErrorMessages(
   errors: BackendError[],
@@ -162,6 +241,14 @@ export function getQueryString(
       .join('&')
   )
 }
+
+export const getTimezone = async (event: H3Event) =>
+  getCookie(event, TIMEZONE_COOKIE_NAME) ||
+  (
+    await ofetch(
+      `http://ip-api.com/json/${event.node.req.headers['x-real-ip']}`
+    )
+  ).timezone
 
 export const isQueryIcFormatValid = (
   ic: LocationQueryValue | LocationQueryValue[]